-w parameter validity check

+++ interfaces/innerkits/app_manager/include/appmgr/app_mgr_client.h
@@ -222,7 +222,7 @@ public:
      *
      * @return Returns ERR_OK on success, others on failure.
      */
-    virtual int FinishUserTest(const std::string &msg, const int &resultCode, const std::string &bundleName);
+    virtual int FinishUserTest(const std::string &msg, const int64_t &resultCode, const std::string &bundleName);
 
     virtual void StartSpecifiedAbility(const AAFwk::Want &want, const AppExecFwk::AbilityInfo &abilityInfo);
