+++ services/abilitymgr/src/ability_manager_stub.cpp
@@ -1227,7 +1227,7 @@ int AbilityManagerStub::StartUserTestInner(MessageParcel &data, MessageParcel &r
 int AbilityManagerStub::FinishUserTestInner(MessageParcel &data, MessageParcel &reply)
 {
     std::string msg = data.ReadString();
-    int resultCode = data.ReadInt32();
+    int64_t resultCode = data.ReadInt64();
     std::string bundleName = data.ReadString();
     int32_t result = FinishUserTest(msg, resultCode, bundleName);
     reply.WriteInt32(result);
